+++ hspec-discover/integration-test/with-no-implicit-prelude/FooSpec.hs
@@ -1,0 +1,13 @@
+module FooSpec (main, spec) where
+
+import           Prelude
+import           Test.Hspec
+
+main :: IO ()
+main = hspec spec
+
+spec :: Spec
+spec = do
+  describe "reverse" $ do
+    it "reverses a list" $ do
+      reverse [1 :: Int, 2, 3] `shouldBe` [3, 2, 1]